+++ lib/buildPackage.nix
@@ -1,7 +1,7 @@
 { cargoBuild
 , installFromCargoBuildLogHook
 , lib
-, removeReferencesTo
+, removeReferencesToVendoredSourcesHook
 }:
 
 { cargoBuildCommand ? "cargoWithProfile build"
@@ -17,12 +17,6 @@ let
   installPhaseCommand = args.installPhaseCommand or ''
     if [ -n "$cargoBuildLog" -a -f "$cargoBuildLog" ]; then
       installFromCargoBuildLog "$out" "$cargoBuildLog"
-
-      # Strip any references to the sources directory (which may have slipped in via
-      # panic info) so depending on the binary doesn't pull in all the sources as well.
-      if [ -z "''${doNotRemoveReferencesToVendorDir-}" ] && [ -n "''${cargoVendorDir-}" ]; then
-        find "$out" -type f -exec remove-references-to -t "$cargoVendorDir" '{}' +
-      fi
     else
       echo ${lib.strings.escapeShellArg ''
         !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
@@ -65,6 +59,6 @@ in
 
   nativeBuildInputs = (old.nativeBuildInputs or [ ]) ++ [
     installFromCargoBuildLogHook
-    removeReferencesTo
+    removeReferencesToVendoredSourcesHook
   ];
 })