+++ pkgs/default.nix
@@ -15,4 +15,5 @@ callPackage:
   inheritCargoArtifactsHook = callPackage ./inheritCargoArtifacts.nix { };
   installCargoArtifactsHook = callPackage ./installCargoArtifacts.nix { };
   installFromCargoBuildLogHook = callPackage ./installFromCargoBuildLog.nix { };
+  removeReferencesToVendoredSourcesHook = callPackage ./removeReferencesToVendoredSources.nix { };
 }