+++ pkgs/removeReferencesToVendoredSources.nix
@@ -1,0 +1,14 @@
+{ makeSetupHook
+, pkgsBuildBuild
+}:
+
+makeSetupHook
+{
+  name = "removeReferencesToVendoredSourcesHook";
+  substitutions = {
+    storeDir = builtins.storeDir;
+  };
+  deps = with pkgsBuildBuild; [
+    removeReferencesTo
+  ];
+} ./removeReferencesToVendoredSourcesHook.sh