+++ CHANGELOG.md
@@ -25,6 +25,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Fixed
 * `cargoAudit` properly keeps any `audit.toml` files when cleaning the source
+* `buildPackage` now has more robust checks to ensure that all references to
+  vendored sources are removed after installation (which avoids consumers of the
+  final binaries having to download the sources as well)
 
 ## [0.6.0] - 2022-09-07