+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## [0.7.0] - 2022-09-28
+
 ## Added
 * `cargoDoc` can now be used for building the documentation of a workspace
 * `cleanCargoSource` can now be used to filter sources to only include cargo and
@@ -207,6 +209,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## 0.1.0 - 2022-01-22
 - First release
 
+[0.7.0]: https://github.com/ipetkov/crane/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/ipetkov/crane/compare/v0.5.1...v0.6.0
 [0.5.1]: https://github.com/ipetkov/crane/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/ipetkov/crane/compare/v0.4.1...v0.5.0